Move Vm to ViewModel folder
