+++ product/src/main/java/com/yas/product/service/ProductService.java
@@ -12,9 +12,9 @@ import com.yas.product.model.ProductImage;
 import com.yas.product.model.ProductOption;
 import com.yas.product.model.ProductOptionCombination;
 import com.yas.product.model.ProductOptionValue;
-import com.yas.product.model.ProductOptionValueSaveVm;
+import com.yas.product.viewmodel.productoption.ProductOptionValueSaveVm;
 import com.yas.product.model.ProductRelated;
-import com.yas.product.model.ProductVariationSaveVm;
+import com.yas.product.viewmodel.product.ProductVariationSaveVm;
 import com.yas.product.model.attribute.ProductAttributeGroup;
 import com.yas.product.model.attribute.ProductAttributeValue;
 import com.yas.product.model.enumeration.FilterExistInWhSelection;
