+++ product/src/main/java/com/yas/product/viewmodel/product/ProductPutVm.java
@@ -2,7 +2,7 @@ package com.yas.product.viewmodel.product;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.yas.product.model.ProductOptionValueSaveVm;
+import com.yas.product.viewmodel.productoption.ProductOptionValueSaveVm;
 import com.yas.product.validation.ValidateProductPrice;
 import jakarta.validation.constraints.NotEmpty;
 import java.util.List;
