+++ product/src/main/java/com/yas/product/viewmodel/product/ProductSaveVm.java
@@ -1,6 +1,5 @@
 package com.yas.product.viewmodel.product;
 
-import com.yas.product.model.ProductVariationSaveVm;
 import java.util.List;
 
 public interface ProductSaveVm<T extends ProductVariationSaveVm> extends ProductProperties {
